# openapi.yaml
...
openapi: "3.2.0"
...
info:
  title: "AniCards Public API"
  version: "1.0.0"
  description: >-
    Hand-maintained OpenAPI 3.2.0 specification for the live AniCards Next.js
    route handlers, validated against the current codebase behavior and tests.
    Storage-related routes persist AniList-derived user snapshots and card
    editor configuration in Upstash Redis; the repository's current privacy
    posture, telemetry consent model, and retention caveats are documented in
    `docs/PRIVACY.md`.
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
  contact:
    name: "Alpha49 (AniCards)"
    url: "https://api.anicards.alpha49.com"
    email: "contact@alpha49.com"
servers:
  - url: "https://api.anicards.alpha49.com"
    description: "Production public API host"
security: []
...
paths:
...
  /api/get-user:
    get:
      operationId: "getStoredUser"
      tags:
        - "user"
      summary: "Retrieve a stored user record"
      description: >-
        Resolves a stored user by numeric `userId` or by normalized `username`.
        When both are provided, `userId` wins. The public response intentionally
        omits internal request metadata such as redacted IP buckets and record
        timestamps.
      parameters:
        - in: query
          name: userId
          required: false
          schema:
            type: integer
            minimum: 1
          description: "Numeric AniList user ID"
        - in: query
          name: username
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 100
          description: >-
            Username lookup key. The server trims and lowercases it before checking
            the username index. Either `userId` or `username` must be supplied.
        - in: query
          name: view
          required: false
          schema:
            type: string
            enum:
              - bootstrap
          description: >-
            Optional lightweight response mode used by the editor bootstrap flow.
            When `view=bootstrap`, the response is reduced to identity fields only.
      responses:
        "200":
          description: "Bounded public user record or lightweight bootstrap identity"
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/PublicUserRecord"
                  - $ref: "#/components/schemas/UserBootstrapRecord"
              examples:
                user:
                  value:
                    userId: 12345
                    username: "sampleuser"
                    stats:
                      User:
                        name: "sampleuser"
                    statistics:
                      anime:
                        count: 420
                        episodesWatched: 5200
                        meanScore: 82
                      manga:
                        count: 90
                        chaptersRead: 1800
                        meanScore: 79
                    favourites:
                      anime:
                        nodes:
                          - id: 1
                            title:
                              romaji: "Fullmetal Alchemist: Brotherhood"
                    pages:
                      followersPage:
                        pageInfo:
                          total: 12
                        followers:
                          - id: 999
                bootstrap:
                  value:
                    userId: 12345
                    username: "sampleuser"
                    avatarUrl: "https://example.com/avatar-medium.webp"
        "400":
          description: "Missing or invalid lookup parameters"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              examples:
                missing:
                  value:
                    error: "Missing userId or username parameter"
                    category: "invalid_data"
                    retryable: false
                    status: 400
                    recoverySuggestions: []
                badUserId:
                  value:
                    error: "Invalid userId parameter"
                    category: "invalid_data"
                    retryable: false
                    status: 400
                    recoverySuggestions: []
                badUsername:
                  value:
                    error: "Invalid username parameter"
                    category: "invalid_data"
                    retryable: false
                    status: 400
                    recoverySuggestions: []
        "404":
          description: "No stored user matched the lookup"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              examples:
                notFound:
                  value:
                    error: "User not found"
                    category: "user_not_found"
                    retryable: false
                    status: 404
                    recoverySuggestions: []
        "500":
          description: "Detected stored-record corruption"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              examples:
                corrupted:
                  value:
                    error: "Stored user record is incomplete or corrupted"
                    category: "server_error"
                    retryable: false
                    status: 500
                    recoverySuggestions: []
        "503":
          description: "User storage temporarily unavailable"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              examples:
                unavailable:
                  value:
                    error: "User data is temporarily unavailable"
                    category: "server_error"
                    retryable: true
                    status: 503
                    recoverySuggestions: []
    options:
      operationId: "preflightGetStoredUser"
      tags:
        - "user"
      summary: "CORS preflight for stored-user lookups"
      description: >-
        Returns the CORS headers required for browser clients to fetch stored
        public user records.
      responses:
        "200":
          $ref: "#/components/responses/JsonGetPreflight"
...
components:
  securitySchemes:
    CronSecret:
      type: apiKey
      in: header
      name: x-cron-secret
      description: >-
        Shared cron secret header for operator-facing cron endpoints. Local and
        manual callers can send `x-cron-secret`. Deployed environments fail
        with `503` when `CRON_SECRET` is missing; unsecured cron is available
        only through the explicit development override.
    CronBearer:
      type: http
      scheme: bearer
      bearerFormat: "CRON_SECRET"
      description: >-
        Hosted schedulers such as Vercel may send `Authorization: Bearer
        <CRON_SECRET>` instead of `x-cron-secret`.
...
  headers:
...
    AccessControlAllowOriginJsonApi:
      description: >-
        Allowed browser origin for JSON API requests. In production this comes
        from `NEXT_PUBLIC_APP_URL`; in development it falls back to the request
        origin or `*`.
      schema:
        type: string
      example: "https://anicards.alpha49.com"
...
    AccessControlAllowHeadersContentType:
      description: "Allowed request headers for this endpoint"
      schema:
        type: string
      example: "Content-Type"
...
    AccessControlAllowMethodsGetHeadOptions:
      description: "Allowed HTTP methods for this endpoint"
      schema:
        type: string
      example: "GET, HEAD, OPTIONS"
...
    AccessControlExposeHeadersJsonApi:
      description: "Response headers exposed to JSON API browser clients"
      schema:
        type: string
      example: "Retry-After, X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset, X-Request-Id"
...
    HeaderVaryOrigin:
      description: "Response varies by request `Origin`."
      schema:
        type: string
      example: "Origin"
    RequestId:
      description: "Per-request correlation ID generated by the API layer"
      schema:
        type: string
      example: "req_01HZQK8F8Y6Z8A2R3M4N5P6Q7R"
  responses:
...
    JsonGetPreflight:
      description: "CORS preflight response for JSON read endpoints"
      headers:
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/AccessControlAllowOriginJsonApi"
        Access-Control-Allow-Headers:
          $ref: "#/components/headers/AccessControlAllowHeadersContentType"
        Access-Control-Allow-Methods:
          $ref: "#/components/headers/AccessControlAllowMethodsGetHeadOptions"
        Access-Control-Expose-Headers:
          $ref: "#/components/headers/AccessControlExposeHeadersJsonApi"
        Vary:
          $ref: "#/components/headers/HeaderVaryOrigin"
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
...
  schemas:
    ApiError:
      type: object
      additionalProperties: true
      properties:
        error:
          type: string
        category:
          type: string
          enum:
            - "user_not_found"
            - "rate_limited"
            - "network_error"
            - "invalid_data"
            - "server_error"
            - "timeout"
            - "authentication"
            - "unknown"
        retryable:
          type: boolean
        status:
          type: integer
          minimum: 400
          maximum: 599
        recoverySuggestions:
          type: array
          items:
            $ref: "#/components/schemas/RecoverySuggestion"
      required:
        - error
        - category
        - retryable
        - status
        - recoverySuggestions
...
    RecoverySuggestion:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
        description:
          type: string
        actionUrl:
          type: string
        actionLabel:
          type: string
      required:
        - title
        - description
...
    PageInfo:
      type: object
      additionalProperties: true
      properties:
        total:
          type: integer
      required:
        - total
    FavouriteCollection:
      type: object
      additionalProperties: true
      properties:
        nodes:
          type: array
          items:
            type: object
            additionalProperties: true
        pageInfo:
          $ref: "#/components/schemas/PageInfo"
...
    UserPages:
      type: object
      additionalProperties: true
      properties:
        followersPage:
          type: object
          additionalProperties: true
        followingPage:
          type: object
          additionalProperties: true
        threadsPage:
          type: object
          additionalProperties: true
        threadCommentsPage:
          type: object
          additionalProperties: true
        reviewsPage:
          type: object
          additionalProperties: true
    UserAggregateKey:
      type: string
      enum:
        - animeSourceMaterialDistributionTotals
        - animeSeasonalPreferenceTotals
        - animeGenreSynergyTotals
        - studioCollaborationTotals
    UserPayloadCompleteness:
      type: object
      additionalProperties: false
      description: >-
        Snapshot completeness metadata describing the sampled user payload that
        was retained in storage.
      properties:
        sampled:
          type: boolean
        fullHistory:
          type: boolean
          enum:
            - false
        boundedSections:
          type: array
          items:
            type: string
        availableAggregates:
          type: array
          items:
            $ref: "#/components/schemas/UserAggregateKey"
        missingAggregates:
          type: array
          items:
            $ref: "#/components/schemas/UserAggregateKey"
      required:
        - sampled
        - fullHistory
        - boundedSections
        - availableAggregates
        - missingAggregates
    UserSnapshotRef:
      type: object
      additionalProperties: false
      description: Stable identifier for one committed user snapshot.
      properties:
        token:
          type: string
        revision:
          type: integer
          minimum: 1
        updatedAt:
          type: string
          format: date-time
        committedAt:
          type: string
          format: date-time
      required:
        - token
        - revision
        - updatedAt
        - committedAt
    PublicUserRecordMetadata:
      type: object
      additionalProperties: false
      description: >-
        Public metadata about the sampled record returned by `/api/get-user`,
        including storage shape and snapshot completeness.
      properties:
        storageFormat:
          type: string
          enum:
            - committed-split
            - legacy-split
            - legacy
        schemaVersion:
          type: integer
          minimum: 1
        snapshot:
          $ref: "#/components/schemas/UserSnapshotRef"
        completeness:
          $ref: "#/components/schemas/UserPayloadCompleteness"
      required:
        - storageFormat
        - schemaVersion
        - completeness
    PublicUserRecord:
      type: object
      additionalProperties: false
      description: >-
        Bounded public lookup DTO returned by `/api/get-user`. Internal
        persistence metadata such as request IP buckets and record timestamps are
        intentionally excluded.
      properties:
        userId:
          type: integer
          minimum: 1
        username:
          type: string
        stats:
          type: object
          additionalProperties: true
          description: "Full stored AniList-derived payload"
        statistics:
          type: object
          additionalProperties: true
          description: "Convenience view of `stats.User.statistics`"
        favourites:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/FavouriteCollection"
        pages:
          $ref: "#/components/schemas/UserPages"
        aggregates:
          type: object
          additionalProperties: true
        recordMeta:
          $ref: "#/components/schemas/PublicUserRecordMetadata"
      required:
        - userId
        - stats
        - statistics
        - favourites
    UserBootstrapRecord:
      type: object
      additionalProperties: false
      description: >-
        Lightweight identity DTO returned by `/api/get-user?view=bootstrap`.
        Used by the editor to resolve canonical usernames and avatars before the
        heavier card payload is loaded.
      properties:
        userId:
          type: integer
          minimum: 1
        username:
          type: string
        avatarUrl:
          type: string
          nullable: true
        recordMeta:
          $ref: "#/components/schemas/PublicUserRecordMetadata"
      required:
        - userId
...